refactor:workbook problem list
