+++ client/src/components/modal/index.tsx
@@ -1,5 +1,5 @@
 import { AiOutlineClose } from 'react-icons/ai';
-import { colors } from '../../styles/theme';
+import { Problem } from '../../types/workbook';
 import CreateProblemModal from './create';
 import SearchProblemModal from './search';
 import { ModalBackground, ModalCloseButton, ModalContainer, ModalInner } from './Style';
@@ -9,23 +9,18 @@ type MODAL_TYPE = 'create' | 'search';
 interface Props {
   type: MODAL_TYPE;
   onToggle: () => void;
+  handleProblemAdd: (problem: Problem) => void;
 }
 
-const ModalContent = {
-  create: <CreateProblemModal />,
-  search: <SearchProblemModal />,
-};
-
-const Modal = ({ type, onToggle }: Props) => {
-  const children = ModalContent[type];
-
+const Modal = ({ type, onToggle, handleProblemAdd }: Props) => {
   return (
     <ModalContainer>
       <ModalInner>
         <ModalCloseButton onClick={onToggle}>
           <AiOutlineClose size={20} />
         </ModalCloseButton>
-        {children}
+        {type === 'create' && <CreateProblemModal handleProblemAdd={handleProblemAdd} />}
+        {type === 'search' && <SearchProblemModal handleProblemAdd={handleProblemAdd} />}
       </ModalInner>
       <ModalBackground onClick={onToggle} />
     </ModalContainer>
