--- client/src/types/question/index.ts
@@ -1,10 +1,0 @@
-interface PostCreateQuestionBody {
-  question: string;
-  questionType: string;
-  answer: string;
-  commentary: string;
-  difficulty: number;
-  hashtags: string[];
-}
-
-export default PostCreateQuestionBody;
